+++ lib/src/clojure_lsp/feature/file_management.clj
@@ -24,9 +24,7 @@
 
 (defn create-ns-changes [uri text db]
   (when-let [new-ns (and (string/blank? text)
-                         (contains? #{:clj :cljs :cljc} (shared/uri->file-type uri))
-                         (not (get (:create-ns-blank-files-denylist db) uri))
-                         (shared/uri->namespace uri db))]
+                         (shared/uri->safe-namespace uri db))]
     (when (settings/get db [:auto-add-ns-to-new-files?] true)
       (let [new-text (format "(ns %s)" new-ns)
             changes [{:text-document {:version (get-in db [:documents uri :v] 0) :uri uri}